chore: rename file to dataItemNames
